BE/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.1.0
pydantic[email]==2.5.0

# Async database support
aiosqlite==0.19.0

# Database drivers (uncomment the one you need)
# PostgreSQL
psycopg2-binary==2.9.9
asyncpg==0.29.0  # For async PostgreSQL - REQUIRED for your setup

# MySQL
# PyMySQL==1.1.0
# mysqlclient==2.2.0  # Alternative MySQL driver
# aiomysql==0.2.0  # For async MySQL

# SQLite (included with Python, aiosqlite for async support)